Initial vue app with main.js replaced by custom code
Ok, this is very react-like so far.

npm init vue@3 - that creates the app. Kinda like create-react-app does.

What's neat to notice is you get a choice of wheather to innclude or not
some core vue libraries for things like routing, testing, formatting etc.

Another new thing is that Vue uses it's own bundling mechanism (vite)
rather than webpack. In similar fashion it also sports similar dedicated
testing toolset - vitest.

How it all starts:

Index.html has a div with an id, and one line below the main.js file is
loaded. It searches for it and mounts the vue up onto it.

It does so using 3 thins:
1. createApp from vue core library (think of of
it as the nail gun).
2. App component which will be nailed into the index page (think of it as the nail)
3. the mount method of crateApp (this tells it where to drive the nail).