Crated and added MainNav component
View is kinda different here from react and angular.
Not only do you have to import a child component,
you also gotta declare it in component configuration object (that would be Vue 2 way if i'm correct)

After that you can use it.

Also the syntax of

{ property }

is ES6 feature where just by naming the property it'll automatically
set it's value to variable of same name as the property.